backend/requirements.txt
# Vaarta Backend Dependencies

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic>=2.7.0,<3

# AI
anthropic==0.40.0
openai==1.54.0
groq>=0.18.0

# PDF
PyMuPDF==1.24.11
reportlab==4.2.2

# Image
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
